+++ src/main/java/org/skypro/starbank/marketing/controller/RecommendationController.java
@@ -1,7 +1,7 @@
 package org.skypro.starbank.marketing.controller;
 
 import org.skypro.starbank.marketing.service.RecommendationService;
-import org.skypro.starbank.marketing.dto.ServiceResult;
+import org.skypro.starbank.marketing.result.ServiceResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
