+++ src/main/java/org/skypro/starbank/marketing/repository/RecommendationsRepository.java
@@ -1,6 +1,6 @@
 package org.skypro.starbank.marketing.repository;
 
-import org.skypro.starbank.marketing.dto.SearchResult;
+import org.skypro.starbank.marketing.result.SearchResult;
 import org.skypro.starbank.marketing.mapper.SearchResultMapper;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
