+++ src/main/java/org/skypro/starbank/marketing/result/SearchResult.java
@@ -1,4 +1,4 @@
-package org.skypro.starbank.marketing.dto;
+package org.skypro.starbank.marketing.result;
 
 public class SearchResult {
     private final Boolean result;
