+++ src/main/java/org/skypro/starbank/marketing/result/ServiceResult.java
@@ -1,6 +1,7 @@
-package org.skypro.starbank.marketing.dto;
+package org.skypro.starbank.marketing.result;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.skypro.starbank.marketing.dto.Recommendation;
 
 import java.util.Collection;
 import java.util.UUID;
